Update settings to use PreferenceFragment
...and also make the settings layout more compliant to the design guide,
which in our case means: no sub screens, only categories.
