+++ Atarashii/src/net/somethingdreadful/MAL/Settings.java
@@ -1,20 +1,10 @@
 package net.somethingdreadful.MAL;
 
 import android.app.ActionBar;
-import android.content.ContentResolver;
-import android.content.Context;
-import android.content.Intent;
-import android.content.SharedPreferences;
-import android.content.SharedPreferences.OnSharedPreferenceChangeListener;
 import android.os.Bundle;
-import android.preference.PreferenceActivity;
-import android.preference.PreferenceManager;
+import android.support.v7.app.ActionBarActivity;
 
-import net.somethingdreadful.MAL.account.AccountService;
-
-public class Settings extends PreferenceActivity implements OnSharedPreferenceChangeListener {
-    Context context;
-    PrefManager Prefs;
+public class Settings extends ActionBarActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -26,36 +16,10 @@ public class Settings extends PreferenceActivity implements OnSharedPreferenceCh
         if (actionBar != null)
             actionBar.setDisplayHomeAsUpEnabled(true);
 
-        context = getApplicationContext();
-        Prefs = new PrefManager(context);
-        PreferenceManager.getDefaultSharedPreferences(this).registerOnSharedPreferenceChangeListener(this);
-
-        addPreferencesFromResource(R.xml.settings);
-
         NfcHelper.disableBeam(this);
-    }
 
-    @Override
-    public void onSharedPreferenceChanged(SharedPreferences sharedPreferences, String key) {
-        String Auth = AccountService.getAuth();
-        Bundle bundle = new Bundle();
-        int interval = Prefs.getSyncTime() * 60;
-        if (key.equals("synchronisation_time")) {
-            ContentResolver.removePeriodicSync(AccountService.getAccount(context), Auth, bundle);
-            ContentResolver.addPeriodicSync(AccountService.getAccount(context), Auth, bundle, interval);
-        } else if (key.equals("synchronisation")) {
-            if (Prefs.getSyncEnabled()) {
-                ContentResolver.setSyncAutomatically(AccountService.getAccount(context), Auth, true);
-                ContentResolver.addPeriodicSync(AccountService.getAccount(context), Auth, bundle, interval);
-            } else {
-                ContentResolver.removePeriodicSync(AccountService.getAccount(context), Auth, bundle);
-                ContentResolver.setSyncAutomatically(AccountService.getAccount(context), Auth, false);
-            }
-        } else if (key.equals("locale")) {
-            sharedPreferences.edit().commit();
-            Intent Home = new Intent(context, Home.class);
-            startActivity(Home);
-            System.exit(0);
-        }
+        getFragmentManager().beginTransaction()
+                .replace(android.R.id.content, new SettingsFragment())
+                .commit();
     }
 }
